Fix a major bug wherein (due to my misunderstanding of the way python passes string arguments) clearing and freeing messages/keys passed into SHA2 and HMAC-SHA2 functions caused that memory to contail NULs when the variable was accessed in python.
Revert to 32bit integer types in the SHA-256 definitions as 64-bit integer types cause digest mismatches vs. OpenSSL's SHA-256. Given that the maximum message size of SHA-256 is 2^64-1, it seems that it should use 64bit integers for the key if not the context, but until I have time to trace all the shifting intricacies and verify vs. NIST test vectors, not gonna mess with it further.

These were both discovered due to the hash output not matching that of Crypto.Hash

Use SHA*_DIGEST_LEN instead of len when zeroing memory for the digest value

Remove usage of Crypto.Hash.{HMAC,SHA256} in utils.py, further reducing dependence on OpenSSL.

+++ cryptoe/src/include/hmac_sha2.h
@@ -67,6 +67,7 @@
 #ifndef HMAC_SHA2_H
 #define HMAC_SHA2_H
 
+#include <inttypes.h>
 #define SHA256_DIGEST_SIZE ( 256 / 8)
 #define SHA384_DIGEST_SIZE ( 384 / 8)
 #define SHA512_DIGEST_SIZE ( 512 / 8)
@@ -77,10 +78,10 @@
 
 
 typedef struct {
-    uint64_t tot_len;
-    uint64_t len;
+    unsigned int tot_len;
+    unsigned int len;
     unsigned char block[2 * SHA256_BLOCK_SIZE];
-    uint32_t h[8];
+    unsigned int h[8];
 } sha256_ctx;
 
 typedef struct {
@@ -133,7 +134,7 @@ void sha384_init(sha384_ctx *ctx);
 void sha512_init(sha512_ctx *ctx);
 
 void sha256_update(sha256_ctx *ctx, const unsigned char *message,
-                   uint64_t len);
+                   unsigned int len);
 void sha384_update(sha384_ctx *ctx, const unsigned char *message,
                    uint64_t len);
 void sha512_update(sha512_ctx *ctx, const unsigned char *message,
@@ -143,7 +144,7 @@ void sha256_final(sha256_ctx *ctx, unsigned char *digest);
 void sha384_final(sha384_ctx *ctx, unsigned char *digest);
 void sha512_final(sha512_ctx *ctx, unsigned char *digest);
 
-void sha256(const unsigned char *message, uint64_t len,
+void sha256(const unsigned char *message, unsigned int len,
             unsigned char *digest);
 void sha384(const unsigned char *message, uint64_t len,
             unsigned char *digest);
